database update incorporating all non-destructive amuzak changes.

+++ lib/class/stream.class.php
@@ -309,7 +309,7 @@ class Stream
     public static function validate_bitrate($bitrate)
     {
         /* Round to standard bitrates */
-        $bit_rate = 16 * (floor($bitrate / 16));
+        $bit_rate = (int) (16 * (floor($bitrate / 16)));
 
         return $bit_rate;
     }
